pyproject.toml
[project]
name = "absense-planner"
description = "Python based absense-planner project built primarily by UK T-Level students."
readme = "README.md"
requires-python = "==3.11.9"
version = "0.0.1"
dependencies = [
  "asgiref==3.7.2",
  "astroid==3.0.1",
  "async-generator==1.10",
  "attrs==23.1.0",
  "beautifulsoup4==4.12.2",
  "behave==1.2.6",
  "black==23.10.1",
  "certifi==2023.7.22",
  "cffi==1.16.0",
  "chardet==5.1.0",
  "charset-normalizer==3.1.0",
  "click==8.1.7",
  "colorama==0.4.6",
  "convertdate==2.4.0",
  "country-list==1.0.0",
  "cryptography==41.0.7",
  "cssselect==1.2.0",
  "dill==0.3.7",
  "django==4.2.6",
  "django-bulma==0.8.3",
  "django-ckeditor==6.7.0",
  "django-codemirror2==0.2",
  "django-cte==1.1.4",
  "django-environ==0.11.2",
  "django-extensions==3.2.3",
  "django-js-asset==2.1.0",
  "django-mptt==0.9.1",
  "django-recurrence",
  "exceptiongroup==1.1.1",
  "execnet==1.9.0",
  "faker==19.12.0",
  "fasteners==0.18",
  "filelock==3.13.0",
  "graphviz==0.20.1",
  "gunicorn==22.0.0",
  "h11==0.14.0",
  "hijri-converter==2.3.1",
  "holidays==0.35",
  "idna==3.4",
  "importlib-metadata==6.8.0",
  "iniconfig==2.0.0",
  "isort==5.12.0",
  "jaraco-classes==3.3.0",
  "keyring==24.2.0",
  "korean-lunar-calendar==0.3.1",
  "lazy-object-proxy==1.9.0",
  "markdown-it-py==3.0.0",
  "mccabe==0.6.0",
  "mdurl==0.1.2",
  "more-itertools==10.1.0",
  "mypy-extensions==1.0.0",
  "names==0.3.0",
  "numpy==1.26.1",
  "order==2.1.4",
  "outcome==1.2.0",
  "packaging==23.2",
  "pandas==2.1.2",
  "parameterized==0.9.0",
  "parse==1.19.1",
  "parse-type==0.6.2",
  "pathspec==0.11.2",
  "pdbp==1.4.2",
  "platformdirs==3.11.0",
  "pluggy==1.2.0",
  "py==1.11.0",
  "pycountry==22.3.5",
  "pycparser==2.21",
  "pydot==1.4.2",
  "pygments==2.15.1",
  "pylint==3.0.2",
  "pylint-django==2.5.5",
  "pylint-plugin-utils==0.8.2",
  "pymeeus==0.5.12",
  "pynose==1.4.5",
  "pyopenssl==23.2.0",
  "pyotp==2.8.0",
  "pyparsing==3.1.1",
  "pyreadline3==3.4.1",
  "pysocks==1.7.1",
  "pytest==7.1.2",
  "pytest-forked==1.6.0",
  "pytest-html==2.0.1",
  "pytest-metadata==3.0.0",
  "pytest-order==1.1.0",
  "pytest-ordering==0.6",
  "pytest-rerunfailures==11.1.2",
  "pytest-xdist==3.3.1",
  "python-dateutil==2.8.2",
  "python-dotenv==1.0.1",
  "pytz==2023.3.post1",
  "pywin32-ctypes==0.2.2",
  "pyyaml==6.0.1",
  "requests==2.31.0",
  "requests-toolbelt==1.0.0",
  "rich==13.4.2",
  "sbvirtualdisplay==1.2.0",
  "scinum==2.0.2",
  "setuptools==69.1.0",
  "six==1.16.0",
  "sniffio==1.3.0",
  "sortedcontainers==2.4.0",
  "soupsieve==2.4.1",
  "sqlparse==0.4.4",
  "tabcompleter==1.2.1",
  "tomli==2.0.1",
  "tomlkit==0.12.1",
  "toposort==1.10",
  "tqdm==4.66.1",
  "trio==0.22.0",
  "trio-websocket==0.10.3",
  "typing-extensions==4.8.0",
  "tzdata==2023.3",
  "urllib3==2.0.7",
  "wrapt==1.15.0",
  "wsproto==1.2.0",
  "zipp==3.17.0",
]

[dependency-groups]
dev = [
  "commitizen",
  "django-debug-toolbar==4.2.0",
  "djhtml",
  "pre-commit",
  "ruff>=0.11.5",
]

[tool.uv]
package = false

[tool.ruff]

[tool.uv.sources]
django-recurrence = { git = "https://github.com/jazzband/django-recurrence.git", rev = "1.12.1" }

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true
